backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-genai==0.2.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
python-dotenv==1.0.1
websockets==13.1
pydantic==2.10.2
pydantic-settings==2.6.1
jinja2==3.1.4
python-multipart==0.0.17
aiofiles==24.1.0
fpdf2==2.8.1
pillow==11.0.0
httpx==0.28.1
